requirements.txt
accelerate==0.23.0
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.0
dataclasses-json==0.6.1
diskcache==5.6.3
filelock==3.12.4
frozenlist==1.4.0
fsspec==2023.9.2
huggingface-hub==0.17.3
idna==3.4
Jinja2==3.1.2
jsonpatch==1.33
jsonpointer==2.4
langchain==0.0.319
langchain-experimental==0.0.32
langsmith==0.0.49
llama_cpp_python==0.2.11
MarkupSafe==2.1.3
marshmallow==3.20.1
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.2
numpy==1.26.1
openai==0.28.1
packaging==23.2
prompt-toolkit==3.0.39
psutil==5.9.6
pydantic==2.4.2
pydantic_core==2.10.1
python-dotenv==1.0.0
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
safetensors==0.4.0
sniffio==1.3.0
SQLAlchemy==2.0.22
sympy==1.12
tenacity==8.2.3
tokenizers==0.14.1
torch==2.1.0
tqdm==4.66.1
transformers==4.34.1
typing-inspect==0.9.0
typing_extensions==4.8.0
urllib3==2.0.7
wcwidth==0.2.8
yarl==1.9.2
